lambda_resgister_access_user/requirements-register.txt
numpy==2.2.5
opencv-python-headless==4.11.0.86
face_recognition==1.2.3
psycopg2-binary
